feat(trace): superpixel + stroke width set in dialog, allow fractional strokes
Numerate's superpixel_width/_height move from "inherited via TraceRenderContext"
to first-class form fields (default 10×10). The transformBeforeSubmit hook and
the Pixelate-grid helperState banner are gone — vectorise.py owns the
superpixel logic now, so the inheritance was vestigial.

Stroke fields go from `int().min(1)` to `number().min(0.1)` with a 0.1
step UI, mirrored in Pydantic + vectorise.py signatures. Lets users draw
hairline outlines on numerate grids and lineart regions.

+++ filter-service/app/vectorise.py
@@ -127,7 +127,7 @@ def grid_lines_svg(
     superpixel_height: int,
     grid_width: int,
     grid_height: int,
-    stroke_width: int,
+    stroke_width: float,
 ) -> list[str]:
     """Vertical + horizontal lines that overlay the cell boundaries."""
     out: list[str] = []
@@ -150,7 +150,7 @@ def numerate_to_svg(
     img: Image.Image,
     superpixel_width: int,
     superpixel_height: int,
-    stroke_width: int,
+    stroke_width: float,
     show_colors: bool,
     num_colors: int = 16,
     on_phase: callable | None = None,
@@ -228,7 +228,7 @@ def numerate_to_svg(
 _FILL_ATTR_RE = re.compile(r'\bfill="(#[0-9A-Fa-f]{6})"')
 
 
-def add_stroke_to_path(path_str: str, color: str, width: int) -> str:
+def add_stroke_to_path(path_str: str, color: str, width: float) -> str:
     """Insert `stroke` + `stroke-width` attributes into a vtracer
     `<path .../>` element. Idempotent: if the path already has a
     stroke, leave it alone."""
@@ -241,7 +241,7 @@ def add_stroke_to_path(path_str: str, color: str, width: int) -> str:
 
 def lineart_to_svg(
     img: Image.Image,
-    line_thickness: int,
+    line_thickness: float,
     blur_amount: int,
     smoothness: float,
     num_colors: int = 8,
